OLMIS-676 Add rsyslog to image
Add rsyslog to image and use custom rsyslog config file. This custom config file enables listening remotely on UDP port and fixes a couple of bugs in original config file (disable loading imklog and change * to :omusrmsg:*).

In order to get rsyslog daemon to run with the Spring Boot app, need to execute both in a shell script.