OLMIS-669 Add support for configuring log level
Add support for configuring log level in a production environment. Configuring the level in a development environment is simple--just modify the configuration file. However, a production environment needs a way to override the configuration file by passing in a configuration file external to the JAR. Provide a way to pass this in.

Add a way to run a docker container in the docker compose file. The docker compose file assumes there is a custom logback.xml file in the parent folder.

Document in the README.